\name{HW3EM-package}
\alias{HW3EM-package}
\alias{HW3EM}
\docType{package}
\title{
\packageTitle{HW3EM}
}
\description{
\packageDescription{HW3EM}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{HW3EM}
\packageIndices{HW3EM}
~~ An overview of how to use the package, including the most ~~
~~ important functions ~~
}
\author{
\packageAuthor{HW3EM}

Maintainer: \packageMaintainer{HW3EM}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file ~~
~~ KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
